#475 Send() in the main file whatsapp.go
